Added proper file handler & sshd log message check
Used with() to new handle for proper file handling

analyse_auth_log() main for loop now checks if "sshd[" is present to
confirm the line is an sshd log message.

analyse_auth_log() now prints a verbose message stdo containing the line
number of any line it encounters that isn't sshd. Later to be revised